Address review: register baseline monitors, move tests out of the monolith
Three fixes from review.

The monitor-cache tests never registered a baseline monitor.
`updateMonitors` only reads and filters `monitors`; `ensureMonitor` is the
only thing that inserts. With the dictionary left empty, `monitorSetChanged`
was true on the very first call, so all three tests exercised the wrong
path and `unchangedMonitorSetKeepsCachedColumnSpans` asserted a value it
could never see. Each test now calls `ensureMonitor` for its baseline
display before resolving a cached span, so the existing-monitor comparison
and the set-changed comparison are both reached as intended.

Per docs/TESTING.md the tests move out of the frozen
`NiriLayoutEngineTests.swift` monolith into a per-behavior file,
`NiriLayoutEngineMonitorCacheTests.swift`. They now use the shared
`makeLayoutPlanTestMonitor` / `layoutPlanTestSyntheticDisplayId` helpers
instead of a monolith-local one, with widths matching the reported
reproduction (2360pt external, 1512pt built-in).

The changeset gains `contributors:` and a `Fixes #197` reference, per
.changeset/README.md.
