+++ .changeset/20260806102735-rescale-column-widths-when-the-connected-display-.md
@@ -1,6 +1,6 @@
 ---
 "nehir": patch
-
+contributors: [lamdor]
 ---
 
-Rescale proportionally sized columns when the connected display changes, so docking or undocking no longer leaves them at the pixel width they had on the previous monitor. Fullscreen and lone windows already filled the new screen and are unaffected.
+Rescale proportionally sized columns when the connected display changes, so docking or undocking no longer leaves them at the pixel width they had on the previous monitor. Fullscreen and lone windows already filled the new screen and are unaffected. Fixes #197
